fix: add service interfaces and implementations
